Changed models & add service layer
